renderer: Split demo codeinto multiple files.
