+++ libopenage/renderer/demo/CMakeLists.txt
@@ -1,5 +1,11 @@
 add_sources(libopenage
+    demo_0.cpp
+    demo_1.cpp
+    demo_2.cpp
+    demo_3.cpp
+    demo_4.cpp
 	tests.cpp
+    util.cpp
 )
 
 pxdgen(
