+++ libopenage/renderer/demo/demo_0.h
@@ -1,0 +1,17 @@
+// Copyright 2015-2023 the openage authors. See copying.md for legal info.
+
+#pragma once
+
+#include "util/path.h"
+
+namespace openage::renderer::tests {
+
+/**
+ * Shows the renderer's ability to create textured renderable objects and
+ * allow basic interaction with them via mouse/key callbacks.
+ *
+ * @param path Path to project rootdir.
+ */
+void renderer_demo_0(const util::Path &path);
+
+} // namespace openage::renderer::tests
