+++ libopenage/renderer/demo/demo_2.h
@@ -1,0 +1,17 @@
+// Copyright 2015-2023 the openage authors. See copying.md for legal info.
+
+#pragma once
+
+#include "util/path.h"
+
+namespace openage::renderer::tests {
+
+/**
+ * Basic test for creating a window and displaying a single-color texture
+ * created by a shader.
+ *
+ * @param path Path to the project rootdir.
+ */
+void renderer_demo_2(const util::Path &path);
+
+} // namespace openage::renderer::tests
