+++ libopenage/renderer/demo/demo_3.h
@@ -1,0 +1,16 @@
+// Copyright 2015-2023 the openage authors. See copying.md for legal info.
+
+#pragma once
+
+#include "util/path.h"
+
+namespace openage::renderer::tests {
+
+/**
+ * Shows the different render stages and camera.
+ *
+ * @param path Path to the project rootdir.
+ */
+void renderer_demo_3(const util::Path &path);
+
+} // namespace openage::renderer::tests
