+++ libopenage/renderer/demo/demo_4.h
@@ -1,0 +1,16 @@
+// Copyright 2015-2023 the openage authors. See copying.md for legal info.
+
+#pragma once
+
+#include "util/path.h"
+
+namespace openage::renderer::tests {
+
+/**
+ * Demo the timing of animation keyframes with the simulation clock.
+ *
+ * @param path Path to the openage asset directory.
+ */
+void renderer_demo_4(const util::Path &path);
+
+} // namespace openage::renderer::tests
