+++ libopenage/renderer/demo/util.cpp
@@ -1,0 +1,8 @@
+// Copyright 2015-2023 the openage authors. See copying.md for legal info.
+
+#include "util.h"
+
+
+namespace openage::renderer::tests {
+
+} // namespace openage::renderer::tests
